+++ backend/internal/routes/routes.go
@@ -5,6 +5,7 @@ import (
 	"github.com/prayudahlah/showflix/backend/internal/title"
 	"github.com/prayudahlah/showflix/backend/internal/login"
 	"github.com/prayudahlah/showflix/backend/internal/dashboard/executive"
+	"github.com/prayudahlah/showflix/backend/internal/dashboard/marketing"
 	"github.com/gofiber/fiber/v2"
 )
 
@@ -19,4 +20,5 @@ func SetupRoutes(app *fiber.App, db *sql.DB) {
 	dash := api.Group("/dashboard")
 
 	executive.RegisterRoutes(dash, db)
+	marketing.RegisterRoutes(dash, db)
 }