+++ frontend/src/api/apiInstance.ts
@@ -1,5 +1,5 @@
 import axios from "axios";
-import camelcaseKeys from 'camelcase-keys';
+import type { ApiError } from "../types/error";
 
 const BASE_URL = import.meta.env.VITE_API_BASE_URL
 
@@ -10,7 +10,6 @@ export const api = axios.create({
 
 api.interceptors.response.use(
   (res) => {
-    res.data = camelcaseKeys(res.data, { deep: true });
     return res
   }, (err) => {
     const status = err.response?.status